Add test for _recover_exposures
